MOBILE-1168 Fix navigation bar color
