+++ Sources/Kevin/Core/Custom/KevinNavigationViewController.swift
@@ -14,8 +14,6 @@ internal class KevinNavigationViewController: UINavigationController {
         super.viewDidLoad()
         
         delegate = self
-        navigationBar.shadowImage = UIImage()
-        navigationBar.isTranslucent = false
         
         interactivePopGestureRecognizer?.delegate = self
         interactivePopGestureRecognizer?.isEnabled = true
@@ -23,29 +21,11 @@ internal class KevinNavigationViewController: UINavigationController {
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        configureNavigationBar()
+        applyKevinNavigationBarStyle()
     }
     
     override var preferredStatusBarStyle: UIStatusBarStyle {
         return .lightContent
-    }
-    
-    private func configureNavigationBar() {
-        if #available(iOS 13.0, *) {
-            let appearance = UINavigationBarAppearance()
-            appearance.configureWithOpaqueBackground()
-            appearance.backgroundColor = UIApplication.shared.isLightThemedInterface ?
-            Kevin.shared.theme.navigationBarStyle.backgroundColorLightMode :
-            Kevin.shared.theme.navigationBarStyle.backgroundColorDarkMode
-            appearance.titleTextAttributes = [NSAttributedString.Key.foregroundColor: Kevin.shared.theme.navigationBarStyle.titleColor]
-            navigationBar.standardAppearance = appearance
-            navigationBar.scrollEdgeAppearance = navigationBar.standardAppearance
-        } else {
-            navigationBar.barTintColor = UIApplication.shared.isLightThemedInterface ?
-            Kevin.shared.theme.navigationBarStyle.backgroundColorLightMode :
-            Kevin.shared.theme.navigationBarStyle.backgroundColorDarkMode
-            navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor: Kevin.shared.theme.navigationBarStyle.titleColor]
-        }
     }
 }
